Split CLI compile helpers
